Changing signature of syscall_smode
All syscalls must use 16-bit arguments, as the A register
(used for passing 8-bit arguments) is used to hold the syscall
ID.

+++ include/syscall.h
@@ -71,7 +71,7 @@ typedef struct _SysInfo_T
 /* Signal handler function pointer type. */
 typedef void (*SIGHANDLER_T)(uint16_t);
 
-void syscall_smode(uint8_t mode);
+void syscall_smode(int mode);
 
 void syscall_dwrite(char * buf, uint32_t sector);
 void syscall_dread(char * buf, uint32_t sector);
